remove: MoreButton 제거
